feat(FX-4027): small refactoring
