+++ src/v2/Apps/Artist/Routes/WorksForSale/Components/ArtistArtworkFilters.tsx
@@ -1,0 +1,45 @@
+import { ColorFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/ColorFilter"
+import { MediumFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/MediumFilter"
+import { PriceRangeFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/PriceRangeFilter"
+import { SizeFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/SizeFilter"
+import { TimePeriodFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/TimePeriodFilter"
+import { WaysToBuyFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/WaysToBuyFilter"
+import { AttributionClassFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/AttributionClassFilter"
+import { ArtworkLocationFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/ArtworkLocationFilter"
+import { ArtistNationalityFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/ArtistNationalityFilter"
+import { MaterialsFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/MaterialsFilter"
+import { PartnersFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/PartnersFilter"
+import { ArtistsFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/ArtistsFilter"
+import { KeywordFilter } from "v2/Components/ArtworkFilter/ArtworkFilters/KeywordFilter"
+import { useFeatureFlag } from "v2/System/useFeatureFlag"
+import type RelayModernEnvironment from "relay-runtime/lib/store/RelayModernEnvironment"
+import { useSystemContext } from "v2/System"
+
+interface ArtistArtworkFiltersProps {
+  relayEnvironment?: RelayModernEnvironment
+}
+
+export const ArtistArtworkFilters: React.FC<ArtistArtworkFiltersProps> = props => {
+  const { relayEnvironment } = props
+  const { user } = useSystemContext()
+
+  const showKeywordFilter = useFeatureFlag("artist-artwork-grid-keyword-search")
+
+  return (
+    <>
+      {showKeywordFilter && <KeywordFilter />}
+      <ArtistsFilter relayEnvironment={relayEnvironment} user={user} expanded />
+      <AttributionClassFilter expanded />
+      <MediumFilter expanded />
+      <PriceRangeFilter expanded />
+      <SizeFilter expanded />
+      <WaysToBuyFilter expanded />
+      <MaterialsFilter />
+      <ArtistNationalityFilter />
+      <ArtworkLocationFilter />
+      <TimePeriodFilter />
+      <ColorFilter />
+      <PartnersFilter />
+    </>
+  )
+}
